pyproject.toml
[project]
name = "image-similarity-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "pillow>=12.2.0",
    "psutil>=7.2.2",
    "sqlite-vec>=0.1.9",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
